==> code/02_erroruncertainty.Rmd <==
---
title: "Error and Uncertainty"
author: "Arfa Aijazi"
date: "April 2022"
output:
  html_document:
    df_print: paged
---
This script completes all calculations related to error and uncertainty on thermal manikin data

### Setup
Load libraries
```{r message=FALSE, warning=FALSE, include=FALSE}
library(tidyverse)
library(lubridate)
library(knitr)
```

Set plot theme and colors
```{r message=FALSE, warning=FALSE, include=FALSE}
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(panel.background = element_rect(colour = "#BBBBBB"))
}
plot_colors <- c("#4477AA", "#CCBB44", "#228833", "#EE6677", "#AA3377", "#66CCEE")
```

### Data processing
Load thermal manikin experimental data
```{r message=FALSE, warning=FALSE, include=FALSE}
manikin_dir <- "../data/experiment/manikin"

files_manikin <- list.files(path = manikin_dir, pattern = "*_Experiment.csv", full.names = TRUE)

colNames_manikin <- read_csv("../data/experiment/manikin/ColumnNames.csv", col_names = F, col_types = cols())

data_manikin <- lapply(files_manikin, read_csv, skip = 5, col_names = colNames_manikin$X1, col_types = cols())
merged_manikin <- reduce(data_manikin, full_join)

data_manikin <- merged_manikin %>%
  select(-ends_with(".Clo"), 
         -ends_with(".Teq"), 
         -ends_with(".PMV"), 
         -ends_with(".PPD"), 
         -ends_with(".SET"), 
         -ends_with(".ET"), 
         -starts_with("All"), 
         -starts_with("GroupA"), 
         -starts_with("GroupB"),
         -Runtime) %>%
  mutate(Time = mdy_hms(Time)) %>%
  drop_na(Time) %>%
  pivot_longer(cols = ends_with(".T") | ends_with(".P"),
               names_to = c("BodySegment", "Metric"),
               names_sep = "[.]") %>%
  pivot_wider(names_from = Metric, names_prefix = "Metric.", values_from = value)

rm(merged_manikin) 
```

Read experimental matrix
```{r message=FALSE, warning=FALSE, include=FALSE}
all_times <- read_csv("../data/experiment/ExperimentalMatrix.csv", col_types = cols()) %>%
  mutate(Start.Time = mdy_hm(Start.Time)) %>%
  mutate(End.Time = mdy_hm(End.Time)) %>%
  mutate(Alias = paste(Chamber.SetPoint, Clothing, Bedding, Posture, Emergency.Blanket, Bed.Type, PCS, Repetition, sep = "_")) %>%
  filter(Skip == FALSE)
```

Subset data to steady state periods. From each experiment, remove the last 1 minute and review preceeding 10 minutes. 
```{r message=FALSE, warning=FALSE, include=FALSE}
all_manikin_data <- data.frame(Alias = character(),
                               Repeated = logical(),
                               Reference = logical(),
                               Chamber.SetPoint = double(),
                               Time = POSIXct(),
                               BodySegment = character(),
                               Metric.T = double(),
                               Metric.P = double())

for (i in 1:nrow(all_times)) {
  subset_manikin <- data_manikin %>%
    filter(between(Time, all_times$End.Time[i]-minutes(11), all_times$End.Time[i]-minutes(1))) %>%
    mutate(Alias = all_times$Alias[i],
           Reference = all_times$Reference[i],
           Repeated.2 = all_times$Repeated.2[i],
           Repeated.3 = all_times$Repeated.3[i],
           Chamber.SetPoint = all_times$Chamber.SetPoint[i])
  
  all_manikin_data <- full_join(all_manikin_data, subset_manikin)
}
```

### Reference cases
Subset reference cases
```{r message=FALSE, warning=FALSE, include=FALSE}
reference_manikin_data_plot <- all_manikin_data %>%
  filter(Reference == TRUE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Timestep = row_number())
```

#### Uncertainty during steady state for a single case
Plot of temperature vs time for reference cases during steady state by body segment
```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(reference_manikin_data_plot, aes(x = Timestep, y = Metric.T, group = Alias)) +
  geom_line(aes(colour = Alias)) +
  theme_custom() +
  theme(legend.position = "none") +
  facet_wrap(~ BodySegment) +
  ggtitle("Reference: Temperature vs. Time")
```

For each reference case and body segment, compute mean temperature during steady state and difference from mean.
```{r echo=TRUE, message=FALSE, warning=FALSE}
reference_manikin_data_T <- all_manikin_data %>%
  filter(Reference == TRUE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Mean.T = mean(Metric.T)) %>%
  mutate(Difference.T = Metric.T - Mean.T)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(reference_manikin_data_T, aes(x = Difference.T)) +
  geom_density(aes(colour = BodySegment), adjust = 5) +
  scale_x_continuous(limits = c(-0.1, 0.1)) +
  theme_custom() +
  ggtitle(label = "Reference: Difference.T within single case during steady state")
```

For each body segment, compute standard deviation and standard uncertainty.
```{r echo=TRUE, message=FALSE, warning=FALSE}
reference_manikin_data_T_sd <- reference_manikin_data_T %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.T = sd(Difference.T)) %>%
  mutate(u = sd.T/sqrt(n)) %>%
  mutate(sd.T = round(sd.T, 2)) %>%
  mutate(u = round(u, 3))
kable(reference_manikin_data_T_sd)
```

Plot of power vs time for reference cases during steady state by body segment\
```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(reference_manikin_data_plot, aes(x = Timestep, y = Metric.P, group = Alias)) +
  geom_line(aes(colour = Alias)) +
  theme_custom() +
  theme(legend.position = "none") +
  facet_wrap(~ BodySegment) +
  ggtitle("Reference: Power vs. Time")
```

For each reference case and body segment, compute mean power during steady state and difference from mean.\
```{r echo=TRUE, message=FALSE, warning=FALSE}
reference_manikin_data_P <- all_manikin_data %>%
  filter(Reference == TRUE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Mean.P = mean(Metric.P)) %>%
  mutate(Difference.P = Metric.P - Mean.P)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(reference_manikin_data_P, aes(x = Difference.P)) +
  geom_density(aes(colour = BodySegment), adjust = 5) +
  scale_x_continuous(limits = c(-2, 2)) +
  theme_custom() +
  ggtitle(label = "Reference: Difference.P within single case during steady state")
```

For each body segment, compute standard deviation and standard uncertainty
```{r echo=TRUE, message=FALSE, warning=FALSE}
reference_manikin_data_P_sd <- reference_manikin_data_P %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.P = sd(Difference.P)) %>%
  mutate(u = sd.P/sqrt(n)) %>%
  mutate(sd.P = round(sd.P, 1)) %>%
  mutate(u = round(u, 2))
kable(reference_manikin_data_P_sd)
```

#### Uncertainty between repeated cases
Compute steady state mean T for each case, then compute mean for groups of repeated cases and difference from group mean for each individual case.
```{r echo=TRUE, message=FALSE, warning=FALSE}
repeated_manikin_data_T <- all_manikin_data %>%
  filter(Reference == TRUE) %>%
  filter(Repeated.3 == TRUE) %>% # filter cases repeated at least 3x
  group_by(Alias, BodySegment) %>%
  summarise(T.SteadyState = mean(Metric.T)) %>%
  mutate(Alias = substr(Alias, 1, nchar(Alias)-2)) %>% # removes repetition number from Alias
  group_by(Alias, BodySegment) %>%
  mutate(T.Mean = mean(T.SteadyState)) %>%
  mutate(Difference.T = T.SteadyState - T.Mean)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(repeated_manikin_data_T, aes(x = Difference.T)) +
  geom_density(aes(colour = BodySegment), adjust = 2) +
  scale_x_continuous(limits = c(-1,1)) +
  theme_custom() +
  ggtitle(label = "Reference: Difference.T between repeated cases during steady state")
```

For each body segment, compute standard deviation and standard uncertainty.
```{r}
repeated_manikin_data_T_sd <- repeated_manikin_data_T %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.T = sd(Difference.T)) %>%
  mutate(u = sd.T/sqrt(n)) %>%
  mutate(sd.T = round(sd.T, 3)) %>%
  mutate(u = round(u, 2))
kable(repeated_manikin_data_T_sd)
```

Compute steady state mean P for each case, then compute mean for groups of repeated cases and difference from group mean for each individual case.
```{r echo=TRUE, message=FALSE, warning=FALSE}
repeated_manikin_data_P <- all_manikin_data %>%
  filter(Reference == TRUE) %>%
  filter(Repeated.3 == TRUE) %>%
  group_by(Alias, BodySegment) %>%
  summarise(P.SteadyState = mean(Metric.P)) %>%
  mutate(Alias = substr(Alias, 1, nchar(Alias)-2)) %>% # removes repetition number from Alias
  group_by(Alias, BodySegment) %>%
  mutate(P.Mean = mean(P.SteadyState)) %>%
  mutate(Difference.P = P.SteadyState - P.Mean)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(repeated_manikin_data_P, aes(x = Difference.P)) +
  geom_density(aes(colour = BodySegment), adjust = 2) +
  scale_x_continuous(limits = c(-10,10)) +
  theme_custom() +
  ggtitle(label = "Reference: Difference.P between repeated cases during steady state")
```

For each body segment, compute standard deviation and standard uncertainty
```{r}
repeated_manikin_data_P_sd <- repeated_manikin_data_P %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.P = sd(Difference.P)) %>%
  mutate(u = sd.P/sqrt(n)) %>%
  mutate(sd.P = round(sd.P, 3)) %>%
  mutate(u = round(u, 2))
kable(repeated_manikin_data_P_sd)
```

### Experimental cases
```{r message=FALSE, warning=FALSE, include=FALSE}
# Subset experimental data
experimental_manikin_data_plot <- all_manikin_data %>%
  filter(Reference == FALSE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Timestep = row_number()) %>%
  filter(Timestep <= 20)
```

#### Uncertainty during steady state for a single case
Plot of temperature vs time for experimental cases during steady state by body segment
```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(experimental_manikin_data_plot, aes(x = Timestep, y = Metric.T, group = Alias)) +
  geom_line(aes(colour = Alias)) +
  theme_custom() +
  theme(legend.position = "none") +
  facet_wrap(~ BodySegment) +
  ggtitle("Experimental: Temperature vs. Time")
```

For each experimental case and body segment, compute mean temperature during steady state and difference from mean.\
```{r echo=TRUE, message=FALSE, warning=FALSE}
experimental_manikin_data_T <- all_manikin_data %>%
  filter(Reference == FALSE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Mean.T = mean(Metric.T)) %>%
  mutate(Difference.T = Metric.T - Mean.T)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(experimental_manikin_data_T, aes(x = Difference.T)) +
  geom_density(aes(colour = BodySegment), adjust = 5) +
  scale_x_continuous(limits = c(-0.1, 0.1)) +
  theme_custom() +
  ggtitle(label = "Experimental: Difference.T within single case during steady state")
```

For each body segment, compute standard deviation and standard uncertainty.
```{r echo=TRUE, message=FALSE, warning=FALSE}
experimental_manikin_data_T_sd <- experimental_manikin_data_T %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.T = sd(Difference.T)) %>%
  mutate(u = sd.T/sqrt(n)) %>%
  mutate(sd.T = round(sd.T, 2)) %>%
  mutate(u = round(u, 3))
kable(experimental_manikin_data_T_sd)
```

Plot of power vs time for reference cases during steady state by body segment
```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(experimental_manikin_data_plot, aes(x = Timestep, y = Metric.P, group = Alias)) +
  geom_line(aes(colour = Alias)) +
  theme_custom() +
  theme(legend.position = "none") +
  facet_wrap(~ BodySegment) +
  ggtitle("Experimental: Power vs. Time")
```

For each experimental case and body segment, compute mean power during steady state and difference from mean.
```{r echo=TRUE, message=FALSE, warning=FALSE}
experimental_manikin_data_P <- all_manikin_data %>%
  filter(Reference == FALSE) %>%
  group_by(Alias, BodySegment) %>%
  mutate(Mean.P = mean(Metric.P)) %>%
  mutate(Difference.P = Metric.P - Mean.P)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(experimental_manikin_data_P, aes(x = Difference.P)) +
  geom_density(aes(colour = BodySegment), adjust = 5) +
  scale_x_continuous(limits = c(-2, 2)) +
  theme_custom() +
  ggtitle(label = "Experimental: Difference.P within single case during steady state")
```

For each body segment, compute standard deviation and standard uncertainty.
```{r echo=TRUE, message=FALSE, warning=FALSE}
experimental_manikin_data_P_sd <- experimental_manikin_data_P %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.P = sd(Difference.P)) %>%
  mutate(u = sd.P/sqrt(n)) %>%
  mutate(sd.P = round(sd.P, 1)) %>%
  mutate(u = round(u, 2))
kable(experimental_manikin_data_P_sd)
```

#### Uncertainty between repeated experimental cases
Compute steady state mean T for each case, then compute mean for groups of repeated cases and difference from group mean for each individual case.
```{r echo=TRUE, message=FALSE, warning=FALSE}
repeated_manikin_data_T <- all_manikin_data %>%
  filter(Reference == FALSE) %>%
  filter(Repeated.3 == TRUE) %>% # filter cases repeated at least 3x
  group_by(Alias, BodySegment) %>%
  summarise(T.SteadyState = mean(Metric.T)) %>%
  mutate(Alias = substr(Alias, 1, nchar(Alias)-2)) %>% # removes repetition number from Alias
  group_by(Alias, BodySegment) %>%
  mutate(T.Mean = mean(T.SteadyState)) %>%
  mutate(Difference.T = T.SteadyState - T.Mean)
```
```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(repeated_manikin_data_T, aes(x = Difference.T)) +
  geom_density(aes(colour = BodySegment), adjust = 2) +
  scale_x_continuous(limits = c(-1,1)) +
  theme_custom() +
  ggtitle(label = "Reference: Difference.T between repeated cases during steady state")
```

For each body segment, compute standard deviation and standard uncertainty.
```{r}
repeated_manikin_data_T_sd <- repeated_manikin_data_T %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.T = sd(Difference.T)) %>%
  mutate(u = sd.T/sqrt(n)) %>%
  mutate(sd.T = round(sd.T, 3)) %>%
  mutate(u = round(u, 2))
kable(repeated_manikin_data_T_sd)
```

Compute steady state mean P for each case, then compute mean for groups of repeated cases and difference from group mean for each individual case.
```{r echo=TRUE, message=FALSE, warning=FALSE}
repeated_manikin_data_P <- all_manikin_data %>%
  filter(Reference == FALSE) %>%
  filter(Repeated.3 == TRUE) %>%
  group_by(Alias, BodySegment) %>%
  summarise(P.SteadyState = mean(Metric.P)) %>%
  mutate(Alias = substr(Alias, 1, nchar(Alias)-2)) %>% # removes repetition number from Alias
  group_by(Alias, BodySegment) %>%
  mutate(P.Mean = mean(P.SteadyState)) %>%
  mutate(Difference.P = P.SteadyState - P.Mean)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(repeated_manikin_data_P, aes(x = Difference.P)) +
  geom_density(aes(colour = BodySegment), adjust = 2) +
  scale_x_continuous(limits = c(-10,10)) +
  theme_custom() +
  ggtitle(label = "Experimental: Difference.P between repeated cases during steady state")
```

For each body segment, compute standard deviation and standard uncertainty
```{r}
repeated_manikin_data_P_sd <- repeated_manikin_data_P %>%
  ungroup() %>%
  group_by(BodySegment) %>%
  summarise(n = n(), sd.P = sd(Difference.P)) %>%
  mutate(u = sd.P/sqrt(n)) %>%
  mutate(sd.P = round(sd.P, 3)) %>%
  mutate(u = round(u, 2))
kable(repeated_manikin_data_P_sd)
```
