==> code/03_h.rmd <==
---
title: "Heat transfer coefficient (h) Calculation"
author: "Arfa Aijazi"
date: "March 2022"
output:
  html_document:
    df_print: paged
---
This script calculates the heat transfer coefficient, h, from reference conditions.

### Setup
Load libraries
```{r message=FALSE, warning=FALSE, include=FALSE}
library(tidyverse)
library(lubridate)
library(knitr)
```

Set plot theme and colors
```{r message=FALSE, warning=FALSE, include=FALSE}
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(panel.background = element_rect(colour = "#BBBBBB"))
}
plot_colors <- c("#4477AA", "#CCBB44", "#228833", "#EE6677", "#AA3377", "#66CCEE")
```

### Data processing
Load thermal maniikin data
```{r message=FALSE, warning=FALSE, include=FALSE}
manikin_dir <- "../data/experiment/manikin"

files_manikin <- list.files(path = manikin_dir, pattern = "*_Experiment.csv", full.names = TRUE)

colNames_manikin <- read_csv("../data/experiment/manikin/ColumnNames.csv", col_names = F, col_types = cols())

data_manikin <- lapply(files_manikin, read_csv, skip = 5, col_names = colNames_manikin$X1, col_types = cols())
merged_manikin <- reduce(data_manikin, full_join)

data_manikin <- merged_manikin %>%
  select(-ends_with(".Clo"), 
         -ends_with(".Teq"), 
         -ends_with(".PMV"), 
         -ends_with(".PPD"), 
         -ends_with(".SET"), 
         -ends_with(".ET"), 
         -starts_with("All"), 
         -starts_with("GroupA"), 
         -starts_with("GroupB"),
         -Runtime) %>%
  mutate(Time = mdy_hms(Time)) %>%
  drop_na(Time) %>%
  pivot_longer(cols = ends_with(".T") | ends_with(".P"),
               names_to = c("BodySegment", "Metric"),
               names_sep = "[.]") %>%
  pivot_wider(names_from = Metric, names_prefix = "Metric.", values_from = value)

rm(merged_manikin) 
```

Load HOBO temperature data
```{r message=FALSE, warning=FALSE, include=FALSE}
hobo_dir <- "../data/experiment/hobo"
merged_hobo <- list()

for (i in 1:4){
  files_hobo <- list.files(path = hobo_dir, pattern = paste0("*_", i, ".csv"), full.names = TRUE)
  
  if(i == 1){
    colNames_hobo <-c("Number", "Time", "Hobo_1.3", "Hobo_1.4", "HostConnected", "Stopped", "EndOfFile")
  }
  
  else{
    colNames_hobo <- c("Number", "Time", paste0("Hobo_",i), "HostConnected", "Stopped", "EndOfFile")
  }
  
  data_hobo <- lapply(files_hobo, read_csv, skip = 2, col_names = colNames_hobo, col_types = cols(.default = col_double(), Time = col_character(), HostConnected = col_character(), Stopped = col_character(), EndOfFile = col_character()))
  merged_hobo[[i]] <-data_hobo %>%
    reduce(full_join) %>%
    select(-Number, -HostConnected, -Stopped, -EndOfFile) %>%
    pivot_longer(cols = starts_with("Hobo_"), names_to = "Hobo", values_to = "Temperature_F")
  
}

data_hobo <- merged_hobo %>%
  reduce(full_join) %>%
  mutate(Temperature_C = (Temperature_F-32)*5/9) %>%
  mutate(Time = mdy_hms(Time)) %>%
  mutate(Hobo = parse_number(Hobo)) %>%
  select(-Temperature_F)

rm(merged_hobo) 
```

Load linear model from HOBO calibration in 01_hobocalibration.Rmd
```{r message=FALSE, warning=FALSE, include=FALSE}
hobo_lm <- read_csv("../data/calibration/HOBO_Calibration.csv", col_types = cols())

hobo_calibrate <- data_hobo %>%
  left_join(hobo_lm) %>%
  mutate(Offset = Intercept + Temperature_C*Slope) %>%
  mutate(Temperature_C_calibrate = Temperature_C - Offset) %>%
  mutate(Temperature_C_calibrate = round(Temperature_C_calibrate, 3)) %>%
  select(Time, Hobo, Temperature_C_calibrate)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Read experimental matrix
experimental_matrix <- read_csv("../data/experiment/ExperimentalMatrix.csv", col_types = cols()) %>%
  mutate(Start.Time = mdy_hm(Start.Time)) %>%
  mutate(End.Time = mdy_hm(End.Time)) %>%
  mutate(Alias = paste(Chamber.SetPoint, Clothing, Bedding, Posture, Emergency.Blanket, Bed.Type, PCS, Repetition, sep = "_"))
```

Read combined uncertainty for manikin temperature and power based on ErrorUncertainty.Rmd
```{r echo=TRUE, message=FALSE, warning=FALSE}
u.reference <- read_csv("../data/uncertainty/uReference.csv", col_types = cols())
```

Calculate h from reference cases with uncertainty propagation
```{r echo=TRUE, message=FALSE, warning=FALSE}
reference <- experimental_matrix %>%
  filter(Reference == TRUE) %>%
  drop_na(Start.Time) %>%
  select(Chamber.SetPoint, Posture, Repetition, End.Time) %>%
  mutate(T.ambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(h.Back = 0) %>%
  mutate(h.Chest = 0) %>%
  mutate(h.Head = 0) %>%
  mutate(h.LFoot = 0) %>%
  mutate(h.LForearm = 0) %>%
  mutate(h.LHand = 0) %>%
  mutate(h.LLowLeg = 0) %>%
  mutate(h.LThigh = 0) %>%
  mutate(h.LUpperArm = 0) %>%
  mutate(h.Pelvis = 0) %>%
  mutate(h.RFoot = 0) %>%
  mutate(h.RForearm = 0) %>%
  mutate(h.RHand = 0) %>%
  mutate(h.RLowLeg = 0) %>%
  mutate(h.RThigh = 0) %>%
  mutate(h.RUpperArm = 0) %>%
  mutate(u.Back = 0) %>%
  mutate(u.Chest = 0) %>%
  mutate(u.Head = 0) %>%
  mutate(u.LFoot = 0) %>%
  mutate(u.LForearm = 0) %>%
  mutate(u.LHand = 0) %>%
  mutate(u.LLowLeg = 0) %>%
  mutate(u.LThigh = 0) %>%
  mutate(u.LUpperArm = 0) %>%
  mutate(u.Pelvis = 0) %>%
  mutate(u.RFoot = 0) %>%
  mutate(u.RForearm = 0) %>%
  mutate(u.RHand = 0) %>%
  mutate(u.RLowLeg = 0) %>%
  mutate(u.RThigh = 0) %>%
  mutate(u.RUpperArm = 0)

for (i in 1:nrow(reference)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, reference$End.Time[i]-minutes(11), reference$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
    filter(between(Temperature_C_calibrate, reference$Chamber.SetPoint[i]-1, reference$Chamber.SetPoint[i]+1)) # drop erroneous temperature measurements (measured temperature is more than +/- 1 deg C from chamber set point)
  reference$T.ambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, reference$End.Time[i]-minutes(11), reference$End.Time[i]-minutes(1)))  %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    mutate(h = Mean.P/(Mean.T - reference$T.ambient[i])) %>%
    left_join(u.reference) %>%
    mutate(bottom = Mean.T - reference$T.ambient[i]) %>%
    mutate(u.bottom = sqrt(reference$u.Tambient[i]^2 + u.T^2)) %>%
    mutate(u = abs(Mean.P/bottom)*sqrt((u.P/Mean.P)^2+(u.bottom/bottom)^2)) %>%
    select(BodySegment, h, u) %>%
    pivot_wider(names_from = BodySegment, values_from = c(h, u), names_sep = ".")
  reference[i,7:38] <- subset_manikin
}
```

Compute mean h by chamber set point and posture with uncertainty propagation
```{r echo=TRUE, message=FALSE, warning=FALSE}
mean.h <- reference %>%
  group_by(Chamber.SetPoint, Posture) %>%
  summarise(across(starts_with("h."), ~mean(.x)), across(starts_with("u."), ~sqrt(sum(.x^2)))) %>%
  select(-u.Tambient) %>%
  pivot_longer(cols = c(starts_with("h."), starts_with("u.")), names_to = c("Metric", "BodySegment"), names_sep = "[.]", values_to = "Value") %>%
  pivot_wider(names_from = Metric, values_from = Value)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
ggplot(mean.h, aes(x = reorder(BodySegment, h), y = h, group = Posture)) +
  geom_errorbar(aes(ymin = h-(2*u), max = h+(2*u)), width = 0.2, position = position_dodge(width = 0.9)) +
  geom_point(aes(colour = Posture, shape = Posture), position = position_dodge(width = 0.9)) + 
  facet_wrap(vars(Chamber.SetPoint)) +
  theme_custom() +
  xlab("") +
  ylab("h (W/m2degC)") +
  coord_flip() 
```

Write h and uncertainty to csv
```{r echo=TRUE, message=FALSE, warning=FALSE}
write_csv(mean.h, "../data/h/h.csv")
```
