==> code/04_Teq.Rmd <==
---
title: "Equivelent Temperature (Teq) Calculation"
author: "Arfa Aijazi"
date: "March 2022"
output:
  html_document:
    df_print: paged
---
This script completes all data analysis related to the calculation of equivlent temperature and the heating/cooling effect

```{r message=FALSE, warning=FALSE, include=FALSE}
# Load libraries
library(tidyverse)
library(lubridate)
library(knitr)
library(ggpubr)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Set plot theme and colors
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(strip.background = element_blank()) +
    theme(panel.border = element_blank()) +
    theme(panel.background = element_blank()) +
    theme(legend.title = element_blank()) +
    theme(axis.title = element_blank()) +
    theme(text = element_text(size = 7, colour = "#000000"))
}
```


```{r message=FALSE, warning=FALSE, include=FALSE}
# Read thermal manikin data
manikin_dir <- "../data/experiment/manikin"

files_manikin <- list.files(path = manikin_dir, pattern = "*_Experiment.csv", full.names = TRUE)

colNames_manikin <- read_csv("../data/experiment/manikin/ColumnNames.csv", col_names = F, col_types = cols())

data_manikin <- lapply(files_manikin, read_csv, skip = 5, col_names = colNames_manikin$X1, col_types = cols())
merged_manikin <- reduce(data_manikin, full_join)

data_manikin <- merged_manikin %>%
  select(-ends_with(".Clo"), 
         -ends_with(".Teq"), 
         -ends_with(".PMV"), 
         -ends_with(".PPD"), 
         -ends_with(".SET"), 
         -ends_with(".ET"), 
         -starts_with("All"), 
         -starts_with("GroupA"), 
         -starts_with("GroupB"),
         -Runtime) %>%
  mutate(Time = mdy_hms(Time)) %>%
  drop_na(Time) %>%
  pivot_longer(cols = ends_with(".T") | ends_with(".P"),
               names_to = c("BodySegment", "Metric"),
               names_sep = "[.]") %>%
  pivot_wider(names_from = Metric, names_prefix = "Metric.", values_from = value)

rm(merged_manikin) 
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Read HOBO raw data
hobo_dir <- "../data/experiment/hobo"
merged_hobo <- list()

for (i in 1:4){
  files_hobo <- list.files(path = hobo_dir, pattern = paste0("*_", i, ".csv"), full.names = TRUE)
  
  if(i == 1){
    colNames_hobo <-c("Number", "Time", "Hobo_1.3", "Hobo_1.4", "HostConnected", "Stopped", "EndOfFile")
  }
  
  else{
    colNames_hobo <- c("Number", "Time", paste0("Hobo_",i), "HostConnected", "Stopped", "EndOfFile")
  }
  
  data_hobo <- lapply(files_hobo, read_csv, skip = 2, col_names = colNames_hobo, col_types = cols(.default = col_double(), Time = col_character(), HostConnected = col_character(), Stopped = col_character(), EndOfFile = col_character()))
  merged_hobo[[i]] <-data_hobo %>%
    reduce(full_join) %>%
    select(-Number, -HostConnected, -Stopped, -EndOfFile) %>%
    pivot_longer(cols = starts_with("Hobo_"), names_to = "Hobo", values_to = "Temperature_F")
  
}

data_hobo <- merged_hobo %>%
  reduce(full_join) %>%
  mutate(Temperature_C = (Temperature_F-32)*5/9) %>%
  mutate(Time = mdy_hms(Time)) %>%
  mutate(Hobo = parse_number(Hobo)) %>%
  select(-Temperature_F)

rm(merged_hobo) 
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# HOBO calibration from hoboCalibration.Rmd

hobo_lm <- read_csv("../data/calibration/HOBO_Calibration.csv", col_types = cols())

hobo_calibrate <- data_hobo %>%
  left_join(hobo_lm) %>%
  mutate(Offset = Intercept + Temperature_C*Slope) %>%
  mutate(Temperature_C_calibrate = Temperature_C - Offset) %>%
  mutate(Temperature_C_calibrate = round(Temperature_C_calibrate, 3)) %>%
  select(Time, Hobo, Temperature_C_calibrate)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Read experimental matrix
experimental_matrix <- read_csv("../data/experiment/ExperimentalMatrix.csv", col_types = cols()) %>%
  mutate(Start.Time = mdy_hm(Start.Time)) %>%
  mutate(End.Time = mdy_hm(End.Time)) %>%
  mutate(Alias = paste(Chamber.SetPoint, Clothing, Bedding, Posture, Emergency.Blanket, Bed.Type, PCS, Repetition, sep = "_"))
```

Read combined uncertainty for manikin temperature and power based on ErrorUncertainty.Rmd
```{r echo=TRUE, message=FALSE, warning=FALSE}
u.experiment <- read_csv("../data/uncertainty/uExperiment.csv", col_types = cols())
```

Load h for each chamber set point and posture from h.Rmd
```{r echo=TRUE, message=FALSE, warning=FALSE}
mean.h <- read_csv("../data/h/h.csv", col_types = cols()) %>%
  pivot_wider(names_from = BodySegment, values_from = c(h, u), names_sep = ".")
```

Load surface area for each body segment
```{r}
SurfaceArea <- read_csv("../data/teq/SurfaceArea.csv", col_types = cols())
```

Calculate Teq from experimental cases with uncertainty propagation
```{r echo=TRUE, message=FALSE, warning=FALSE}
experiment <- experimental_matrix %>%
  filter(Reference == FALSE) %>%
  filter(Chamber.SetPoint != 24) %>%
  drop_na(Start.Time) %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  left_join(mean.h) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Teq = 0) %>%
  mutate(u.Teq = 0) %>%
  mutate(T.skin = 0) %>%
  mutate(Q = 0) %>%
  mutate(h.cal = 0) %>%
  filter(Skip == FALSE)

for (i in 1:nrow(experiment)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, experiment$End.Time[i]-minutes(11), experiment$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, experiment$Chamber.SetPoint[i]-2, experiment$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  experiment$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_h <- experiment[i,20:51] %>%
    pivot_longer(cols = c(starts_with("h."), starts_with("u.")), names_to = c("Metric", "BodySegment"), values_to = "Value", names_sep = "[.]") %>%
    pivot_wider(names_from = Metric, values_from = Value) %>%
    rename(Mean.h = h, u.h = u)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, experiment$End.Time[i]-minutes(11), experiment$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    left_join(subset_h) %>%
    mutate(Poverh = Mean.P/Mean.h) %>%
    mutate(u.Poverh = abs(Mean.P/Mean.h)*sqrt((u.P/Mean.P)^2+(u.h/Mean.h)^2)) %>%
    mutate(Teq = Mean.T - Poverh) %>%
    mutate(uc = sqrt(u.T^2 + u.Poverh^2)) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  experiment$Teq[i] <- sum(subset_manikin$Teq*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  experiment$u.Teq[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$uc)^2))
  experiment$T.skin[i] <- sum(subset_manikin$Mean.T*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  experiment$Q[i] <- sum(subset_manikin$Mean.P*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  experiment$h.cal[i] <- sum(subset_manikin$Mean.h*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  
  experiment <- experiment %>%
    mutate(h.body = Q/(T.skin - Tambient))
}
```
Subset baseline cases
```{r echo=TRUE, message=FALSE, warning=FALSE}
baseline <- experiment %>%
  filter(Baseline == TRUE) %>%
  group_by(Chamber.SetPoint) %>%
  summarise(Baseline.Teq = mean(Teq), Baseline.u.Teq = sqrt(sum(u.Teq^2)), T.skin_Baseline = mean(T.skin), Q_Baseline = mean(Q), h.cal_Baseline = mean(h.cal), h.body_Baseline = mean(h.body))
```

Calculated thermal effect
```{r echo=TRUE, message=FALSE, warning=FALSE}
thermal.effect <- experiment %>%
  filter(Baseline == FALSE) %>%
  select(1:10, Alias, Tambient, u.Tambient, Teq, u.Teq, T.skin, h.body, h.cal) %>%
  left_join(baseline) %>%
  mutate(ThermalEffect = Teq - Baseline.Teq) %>%
  mutate(u.ThermalEffect = sqrt(u.Teq^2 + Baseline.u.Teq^2)) 

thermal.effect_group <- thermal.effect %>%
  mutate(Alias = substr(Alias, 1, nchar(Alias)-2)) %>% # removes repetition number from Alias
  group_by(Alias) %>%
  summarise(ThermalEffect = mean(ThermalEffect), u.ThermalEffect = sqrt(sum(u.ThermalEffect^2)), T.skin = mean(T.skin), h.body = mean(h.body), h.cal = mean(h.cal)) %>%
  ungroup() %>%
  mutate(Alias2 = Alias) %>%
  separate(Alias2, c("Chamber.SetPoint", "Clothing", "Bedding", "Posture", "EmergencyBlanket", "Bed", "PCS", "Repetition")) %>%
  mutate(Chamber.SetPoint = as.numeric(Chamber.SetPoint)) %>%
  mutate(Clothing = factor(Clothing, levels = c("Nude", "Light", "Heavy"))) %>%
  mutate(Bedding = factor(Bedding, levels = c("None", "Light", "Heavy"))) %>%
  mutate(Posture = factor(Posture, levels = c("Starfish", "Log", "Fetal"))) %>%
  mutate(EmergencyBlanket = factor(EmergencyBlanket, levels = c("FALSE", "TRUE"))) %>%
  mutate(Bed = factor(Bed, levels = c("Slat", "Mattress"))) %>%
  mutate(System = case_when(PCS == "None" ~ "Passive",
                            PCS != "None" & Chamber.SetPoint == 16 ~ "Active: Heating",
                            PCS != "None" & Chamber.SetPoint == 28 ~ "Active: Cooling")) %>%
  mutate(System = factor(System, levels = c("Active: Cooling", "Passive", "Active: Heating"))) %>%
  select(-Repetition) %>%
  mutate(Abs.ThermalEffect = abs(ThermalEffect))
```

```{r}
thermal.effect_labels <- read_csv("../data/teq/thermal.effect_labels.csv")

thermal.effect_plot <- thermal.effect_group %>%
  left_join(thermal.effect_labels)
```
```{r}
k = 2

# Full Heating plot
thermal.effect_plot_heating <- thermal.effect_plot %>%
  filter(Chamber.SetPoint == 16) %>%
  mutate(System = case_when(PCS == "None" ~ "Passive",
                            PCS != "None" & Clothing == "Light" ~ "Active",
                            PCS != "None" & Clothing != "Light" ~ "Passive + Active")) %>%
  mutate(System = factor(System, levels = c("Passive", "Active", "Passive + Active"))) %>%
  arrange(System, desc(Abs.ThermalEffect)) %>%
  mutate(Order = row_number())

heating_plot <- ggplot(thermal.effect_plot_heating, aes(x = reorder(Label, desc(Order)), y = Abs.ThermalEffect)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  geom_hline(yintercept = 5.9778057, linetype = "dotted") +
  geom_errorbar(aes(ymin = Abs.ThermalEffect - k*u.ThermalEffect, ymax = Abs.ThermalEffect + k*u.ThermalEffect), width = 0.2) +
  geom_point(aes(colour = System, shape = System), size = 2) +
  theme_custom() +
  scale_y_continuous(breaks = seq(0, 7, by = 1), limits = c(-.31,7.41)) +
  scale_colour_manual(values = c("#ee833d", "#59abb7", "#495893")) +
  scale_x_discrete(labels = function(Label) str_wrap(Label, width = 35)) +
  theme(plot.margin = margin(t = 0, r = 0, b = 0, l = 0)) +
  theme(axis.title.x = element_text()) +
  #theme(axis.line.x = element_line(), axis.ticks.x = element_line()) +
  theme(axis.text.x = element_blank()) +
  coord_flip()
```

```{r}
# Cooling plot
thermal.effect_plot_cooling <- thermal.effect_plot %>%
  filter(Chamber.SetPoint == 28) %>%
  mutate(System = case_when(PCS == "None" ~ "Passive",
                            PCS != "None" & Clothing == "Light" ~ "Active",
                            PCS != "None" & Clothing != "Light" ~ "Passive + Active")) %>%
  mutate(System = factor(System, levels = c("Passive", "Active", "Passive + Active"))) %>%
  arrange(System, desc(Abs.ThermalEffect)) %>%
  mutate(Order = row_number())

cooling_plot <- ggplot(thermal.effect_plot_cooling, aes(x = reorder(Label, desc(Order)), y = Abs.ThermalEffect)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  geom_hline(yintercept = 2.51529887, linetype = "dotted") +
  geom_errorbar(aes(ymin = Abs.ThermalEffect - k*u.ThermalEffect, ymax = Abs.ThermalEffect + k*u.ThermalEffect), width = 0.2) +
  geom_point(aes(colour = System, shape = System), size = 2) +
  theme_custom() +
  scale_y_continuous(breaks = seq(0, 7, by = 1), limits = c(-.31,7.41)) +
  scale_colour_manual(values = c("#ee833d", "#59abb7", "#495893")) +
  scale_x_discrete(labels = function(Label) str_wrap(Label, width = 35)) +
  theme(plot.margin = margin(t = 0, r = 0, b = 0, l = 0)) +
  theme(axis.title.x = element_text()) +
  theme(axis.line.x = element_line(), axis.ticks.x = element_line()) +
  coord_flip()
  
```

```{r}
# Combined full plot

ggarrange(heating_plot, cooling_plot, common.legend = TRUE, ncol = 1, nrow = 2, align = "hv")

# ggsave("thermal.effect_plot.pdf", width = 6.5, height = 5, units = "in")
```
