==> code/06_fanspeed.Rmd <==
---
title: "Fan Speed Calculation"
author: "Arfa Aijazi"
date: "April 2022"
output:
  html_document:
    df_print: paged
---
This script plots the spatial variation in air speed due to the ceiling and fan pedestal fan as well as calculates the mean air speed over the bed.

```{r message=FALSE, warning=FALSE, include=FALSE}
# Load libraries
library(tidyverse)
library(lubridate)
library(knitr)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Set plot theme and colors
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(panel.background = element_rect(colour = "#BBBBBB")) +
    theme(aspect.ratio = 1.5) +
    theme(axis.title = element_blank(), axis.text = element_blank())
}
plot_colors <- c("#4477AA", "#CCBB44", "#228833", "#EE6677", "#AA3377", "#66CCEE")
```

```{r}
FanData <- read_csv("../data/fan_speed/FanSpeed.csv") %>%
  mutate(AirSpeed = round(AirSpeed, 1))
```
```{r}
ggplot(FanData, aes(x = Row, y = Column, fill = AirSpeed)) +
  geom_tile() + 
  facet_grid(Setting~PCS) +
  theme_custom() +
  scale_fill_gradient(low = "#E7D4E8",
                      high = "#762A83",
                      limits = c(0, 1.1),
                      breaks = c(0, 0.5, 1))

# ggsave("fanspeedplot.pdf", width = 4, height = 6, units = "in")
```

```{r}
MeanFanSpeed <- FanData %>%
  group_by(PCS, Setting) %>%
  summarise(Mean = mean(AirSpeed)) %>%
  mutate(Mean = round(Mean, 1))

MeanFanSpeed
```
