==> code/07_casestudy_map.Rmd <==
---
title: "Case Study Map"
author: "Arfa Aijazi"
date: "April 2022"
output:
  html_document:
    df_print: paged
---
This script produces a map of the two case study locations

```{r}
library(maps)
library(tidyverse)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Set plot theme and colors
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(strip.background = element_blank()) +
    theme(panel.border = element_blank()) +
    theme(panel.background = element_blank()) +
    theme(legend.title = element_blank()) +
    theme(text = element_text(size = 7, colour = "#000000")) +
    theme(plot.margin = margin(t = 0.1, r = 0, b = 0.1, l = 0))
}
```

```{r}

CaseStudy <- read_csv("../data/case_study/CaseStudyLatLong.csv")

states_map <- map_data("state")
world_map <- map_data("world")

world_map <- world_map %>%
  filter(region != "Antarctica")

world_map$region <- ifelse(world_map$subregion == "Alaska", "Alaska", world_map$region)

ggplot() + 
  geom_map(aes(map_id = region), 
           map = world_map, 
           data = world_map, 
           color = "#bbbbbb",
           fill = "#ffffff",
           size = 0.2) + 
  geom_map(aes(map_id = region), 
           map = states_map, 
           data = states_map, 
           color = "#bbbbbb",
           fill = "#ffffff",
           size = 0.2) + 
  geom_point(aes(x = CaseStudy$Longitude, y = CaseStudy$Latitude, shape = CaseStudy$Location, colour = CaseStudy$Location), size = 2) +
  expand_limits(x = world_map$long, y = world_map$lat) +
  theme_custom() +
  theme(legend.position = "none") +
  theme(axis.text = element_blank()) +
  theme(axis.title = element_blank()) + 
  coord_fixed(1.3)

# ggsave("CaseStudy_map.pdf", width = 4, height = 3, units = "in")

```
