==> code/08_clo.Rmd <==
---
title: "Clo"
author: "Arfa Aijazi"
date: "May 2022"
output: html_document
---
This script calculates the clo value of clothing and bedding combinations considered in the study

```{r message=FALSE, warning=FALSE, include=FALSE}
# Load libraries
library(tidyverse)
library(lubridate)
library(knitr)
library(ggpubr)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Set plot theme and colors
theme_custom = function() {
  theme_minimal() %+replace%
    theme(legend.position = "top") +
    theme(panel.grid = element_blank()) +
    theme(strip.background = element_blank()) +
    theme(panel.border = element_blank()) +
    theme(panel.background = element_blank()) +
    theme(legend.title = element_blank()) +
    theme(axis.title = element_blank()) +
    theme(text = element_text(size = 7, colour = "#000000"))
}
```


```{r message=FALSE, warning=FALSE, include=FALSE}
# Read thermal manikin data
manikin_dir <- "../data/experiment/manikin"

files_manikin <- list.files(path = manikin_dir, pattern = "*_Experiment.csv", full.names = TRUE)

colNames_manikin <- read_csv("../data/experiment/manikin/ColumnNames.csv", col_names = F, col_types = cols())

data_manikin <- lapply(files_manikin, read_csv, skip = 5, col_names = colNames_manikin$X1, col_types = cols())
merged_manikin <- reduce(data_manikin, full_join)

data_manikin <- merged_manikin %>%
  select(-ends_with(".Clo"), 
         -ends_with(".Teq"), 
         -ends_with(".PMV"), 
         -ends_with(".PPD"), 
         -ends_with(".SET"), 
         -ends_with(".ET"), 
         -starts_with("All"), 
         -starts_with("GroupA"), 
         -starts_with("GroupB"),
         -Runtime) %>%
  mutate(Time = mdy_hms(Time)) %>%
  drop_na(Time) %>%
  pivot_longer(cols = ends_with(".T") | ends_with(".P"),
               names_to = c("BodySegment", "Metric"),
               names_sep = "[.]") %>%
  pivot_wider(names_from = Metric, names_prefix = "Metric.", values_from = value)

rm(merged_manikin) 
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Read HOBO raw data
hobo_dir <- "../data/experiment/hobo"
merged_hobo <- list()

for (i in 1:4){
  files_hobo <- list.files(path = hobo_dir, pattern = paste0("*_", i, ".csv"), full.names = TRUE)
  
  if(i == 1){
    colNames_hobo <-c("Number", "Time", "Hobo_1.3", "Hobo_1.4", "HostConnected", "Stopped", "EndOfFile")
  }
  
  else{
    colNames_hobo <- c("Number", "Time", paste0("Hobo_",i), "HostConnected", "Stopped", "EndOfFile")
  }
  
  data_hobo <- lapply(files_hobo, read_csv, skip = 2, col_names = colNames_hobo, col_types = cols(.default = col_double(), Time = col_character(), HostConnected = col_character(), Stopped = col_character(), EndOfFile = col_character()))
  merged_hobo[[i]] <-data_hobo %>%
    reduce(full_join) %>%
    select(-Number, -HostConnected, -Stopped, -EndOfFile) %>%
    pivot_longer(cols = starts_with("Hobo_"), names_to = "Hobo", values_to = "Temperature_F")
  
}

data_hobo <- merged_hobo %>%
  reduce(full_join) %>%
  mutate(Temperature_C = (Temperature_F-32)*5/9) %>%
  mutate(Time = mdy_hms(Time)) %>%
  mutate(Hobo = parse_number(Hobo)) %>%
  select(-Temperature_F)

rm(merged_hobo) 
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# HOBO calibration from hoboCalibration.Rmd

hobo_lm <- read_csv("../data/calibration/HOBO_Calibration.csv", col_types = cols())

hobo_calibrate <- data_hobo %>%
  left_join(hobo_lm) %>%
  mutate(Offset = Intercept + Temperature_C*Slope) %>%
  mutate(Temperature_C_calibrate = Temperature_C - Offset) %>%
  mutate(Temperature_C_calibrate = round(Temperature_C_calibrate, 3)) %>%
  select(Time, Hobo, Temperature_C_calibrate)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Read experimental matrix
experimental_matrix <- read_csv("../data/experiment/ExperimentalMatrix.csv", col_types = cols()) %>%
  mutate(Start.Time = mdy_hm(Start.Time)) %>%
  mutate(End.Time = mdy_hm(End.Time)) %>%
  mutate(Alias = paste(Chamber.SetPoint, Clothing, Bedding, Posture, Emergency.Blanket, Bed.Type, PCS, Repetition, sep = "_"))
```

Read combined uncertainty for manikin temperature and power based on ErrorUncertainty.Rmd
```{r echo=TRUE, message=FALSE, warning=FALSE}
u.experiment <- read_csv("../data/uncertainty/uExperiment.csv", col_types = cols())
```

```{r}
mattress_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

slat_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Slat", PCS == "None") %>%
  mutate(Type = "Baseline")

mattress_allpassive <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Nude", Bedding == "None", Posture == "Starfish", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "AllPassive")

slat_allpassive <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Nude", Bedding == "None", Posture == "Starfish", Emergency.Blanket == FALSE, Bed.Type == "Slat", PCS == "None") %>%
  mutate(Type = "AllPassive")

mattress_clo <- mattress_baseline %>%
  full_join(slat_baseline) %>%
  full_join(mattress_allpassive) %>%
  full_join(slat_allpassive)
```
Load surface area for each body segment
```{r}
SurfaceArea <- read_csv("../data/teq/SurfaceArea.csv", col_types = cols())
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
mattress <- mattress_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(mattress_clo)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, mattress$End.Time[i]-minutes(11), mattress$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, mattress$Chamber.SetPoint[i]-2, mattress$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  mattress$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, mattress$End.Time[i]-minutes(11), mattress$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - mattress$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + mattress$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  mattress$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  mattress$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```
Averages
```{r}
mattress_avg <- mattress %>%
  group_by(Bed.Type, Type) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = Bed.Type, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_Mattress - Iclo_Slat) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_Mattress^2+u.Iclo_Slat^2))

mattress_clo <- mean(mattress_avg$Iclo_Difference)
```
```{r}
sheet_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

nosheet_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Light", Bedding == "None", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

sheet_clo <- sheet_baseline %>%
  full_join(nosheet_baseline)
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
sheet <- sheet_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(sheet)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, sheet$End.Time[i]-minutes(11), sheet$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, sheet$Chamber.SetPoint[i]-2, sheet$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  sheet$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, sheet$End.Time[i]-minutes(11), sheet$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - sheet$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + sheet$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  sheet$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  sheet$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```
Averages
```{r}
sheet_avg <- sheet %>%
  group_by(Bedding) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = Bedding, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_Light - Iclo_None) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_Light^2+u.Iclo_None^2))

sheet_clo <- mean(sheet_avg$Iclo_Difference)
```

```{r}
blanket_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Light", Bedding == "Heavy", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

sheet_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

blanket_clo <- blanket_baseline %>%
  full_join(sheet_baseline)
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
blanket <- blanket_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(blanket)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, blanket$End.Time[i]-minutes(11), blanket$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, blanket$Chamber.SetPoint[i]-2, blanket$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  blanket$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, blanket$End.Time[i]-minutes(11), blanket$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - blanket$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + blanket$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  blanket$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  blanket$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```

Averages
```{r}
blanket_avg <- blanket %>%
  group_by(Bedding, Type) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = Bedding, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_Heavy - Iclo_Light) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_Heavy^2+u.Iclo_Light^2))

blanket_clo <- mean(blanket_avg$Iclo_Difference)
```

```{r}
lightclo_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

nude_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 28, Clothing == "Nude", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

lightclo_clo <- lightclo_baseline %>%
  full_join(nude_baseline)
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
lightclo <- lightclo_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(lightclo)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, lightclo$End.Time[i]-minutes(11), lightclo$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, lightclo$Chamber.SetPoint[i]-2, lightclo$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  lightclo$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, lightclo$End.Time[i]-minutes(11), lightclo$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - lightclo$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + lightclo$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  lightclo$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  lightclo$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```

```{r}
lightclo_avg <- lightclo %>%
  group_by(Clothing, Type) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = Clothing, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_Light - Iclo_Nude) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_Light^2+u.Iclo_Nude^2))

lightclo_clo <- mean(lightclo_avg$Iclo_Difference)
```

```{r}
heavyclo_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Heavy", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

lightclo_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

heavyclo_clo <- heavyclo_baseline %>%
  full_join(lightclo_baseline)
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
heavyclo <- heavyclo_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(heavyclo)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, heavyclo$End.Time[i]-minutes(11), heavyclo$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, heavyclo$Chamber.SetPoint[i]-2, heavyclo$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  heavyclo$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, heavyclo$End.Time[i]-minutes(11), heavyclo$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - heavyclo$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + heavyclo$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  heavyclo$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  heavyclo$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```

```{r}
heavyclo_avg <- heavyclo %>%
  group_by(Clothing, Type) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = Clothing, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_Heavy - Iclo_Light) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_Heavy^2+u.Iclo_Light^2))

heavyclo_clo <- mean(heavyclo_avg$Iclo_Difference)
```

```{r}
heatedblanket_baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "HeatedBlanketOff") %>%
  mutate(Type = "Baseline")

baseline <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Light", Bedding == "Light", Posture == "Log", Emergency.Blanket == FALSE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "Baseline")

heatedblanket_allpassive <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Heavy", Bedding == "Heavy", Posture == "Fetal", Emergency.Blanket == TRUE, Bed.Type == "Mattress", PCS == "HeatedBlanketOff") %>%
  mutate(Type = "AllPassive")

allpassive <- experimental_matrix %>%
  filter(Chamber.SetPoint == 16, Clothing == "Heavy", Bedding == "Heavy", Posture == "Fetal", Emergency.Blanket == TRUE, Bed.Type == "Mattress", PCS == "None") %>%
  mutate(Type = "AllPassive")


heatedblanket_clo <- heatedblanket_baseline %>%
  full_join(baseline)
```

```{r echo=TRUE, message=FALSE, warning=FALSE}
heatedblanket <- heatedblanket_clo %>%
  select(-Random.Number, -Reference, Notes, Comments, IR.Image) %>%
  mutate(Tambient = 0) %>%
  mutate(u.Tambient = 0.068) %>%
  mutate(Clo = 0) %>%
  mutate(u.Clo = 0)

for (i in 1:nrow(heatedblanket)) {
  subset_hobo <- hobo_calibrate %>%
    filter(between(Time, heatedblanket$End.Time[i]-minutes(11), heatedblanket$End.Time[i]-minutes(1))) %>% # drop last 1 minute and average over the previous 10 minutes
     filter(between(Temperature_C_calibrate, heatedblanket$Chamber.SetPoint[i]-2, heatedblanket$Chamber.SetPoint[i]+2)) # drop erroneous temperature measurements (measured temperature is more than +/- 2 deg C from chamber set point)
  heatedblanket$Tambient[i] <- mean(subset_hobo$Temperature_C_calibrate)
  
  subset_manikin <- data_manikin %>%
    filter(between(Time, heatedblanket$End.Time[i]-minutes(11), heatedblanket$End.Time[i]-minutes(1))) %>%
    group_by(BodySegment) %>%
    summarise(Mean.T = mean(Metric.T), Mean.P = mean(Metric.P)) %>%
    left_join(u.experiment) %>%
    mutate(deltaT = Mean.T - heatedblanket$Tambient[i]) %>%
    mutate(u.deltaT = sqrt(u.T^2 + heatedblanket$u.Tambient[i]^2)) %>%
    mutate(Iclo = deltaT/(Mean.P*0.155)) %>%
    mutate(u.Iclo = abs(deltaT/(Mean.P*0.155)*sqrt((u.deltaT/deltaT)^2+((u.P*0.155)/(Mean.P*0.155)^2)))) %>%
    left_join(SurfaceArea) %>%
    mutate(weight = SurfaceArea/sum(SurfaceArea))
  heatedblanket$Clo[i] <- sum(subset_manikin$Iclo*subset_manikin$SurfaceArea)/sum(subset_manikin$SurfaceArea)
  heatedblanket$u.Clo[i] <- sqrt(sum((subset_manikin$weight*subset_manikin$u.Iclo)^2))
}
```

```{r}
heatedblanket_avg <- heatedblanket %>%
  group_by(PCS, Type) %>%
  summarise(Iclo = mean(Clo), u.Iclo = sqrt(sum(u.Clo^2))) %>%
  pivot_wider(names_from = PCS, values_from = c(Iclo, u.Iclo)) %>%
  mutate(Iclo_Difference = Iclo_HeatedBlanketOff - Iclo_None) %>%
  mutate(u.Iclo_Difference = sqrt(u.Iclo_HeatedBlanketOff^2+u.Iclo_None^2))

heatedblanket_clo <- mean(heatedblanket_avg$Iclo_Difference)
```
